refactor: move wallet not available component to itWallet feature
